First pass at upgrading rails.
This was the result of:
* updating the rails version in the Gemfile.
* Running app update to clobber all config files.
* Fixing issues from seeding and clicking around

The `auto_html_for` sections appear to deadlock the app. They are
commented out for now.

Other change was opinionated to get parity between dev and prod db &
webserver. I.e in dev use pg and puma over mysql and unicorn.

There does not appear to be tests so I didn't bother with verifying
anything works there.